add: milvus rework; milvus client integration into image repo; ids type fix

+++ backend/internal/model/product.go
@@ -1,11 +1,11 @@
 package model
 
 type Product struct {
-	ID          int     `db:"id"`
+	ID          int64   `db:"id"`
 	Name        string  `db:"name"`
 	Description string  `db:"description"`
 	Price       float32 `db:"price"`
-	CategoryID  int     `db:"category_id"`
+	CategoryID  int64   `db:"category_id"`
 }
 
 // Добавить конструкторы?
